KS-2-1 average score

+++ src/main/kotlin/lesson2/lesson2_task1.kt
@@ -1,12 +1,12 @@
 package org.example.lesson2
 
 fun main() {
-    val englishGradeofStudent1 = 3
-    val englishGradeofStudent2 = 4
-    val englishGradeofStudent3 = 3
-    val englishGradeofStudent4 = 5
+    val englishGradeofStudent1 = (3).toFloat()
+    val englishGradeofStudent2 = (4).toFloat()
+    val englishGradeofStudent3 = (3).toFloat()
+    val englishGradeofStudent4 = (5).toFloat()
 
-    val numberOfStudents = 4f
+    val numberOfStudents = 4
 
     val averageScoreInEnglish = (
         englishGradeofStudent1 + englishGradeofStudent2 +
